Add action to force resync of images in all pools on local endpoint
There exist failure scenarios where abrupt shutdown and/or
interruptions to communication may lead to a split-brain situation
where the RBD Mirroring process in both Ceph clusters claim to be
the primary.

In such a situation the operator must decide which cluster has the
most recent data and should be elected primary by using the
``demote`` and ``promote`` (optionally with force parameter)
actions.

After making this decision the secondary cluster must be resynced
to track the promoted master, this is done by running the
``resync-pools`` action on the non-master cluster.

Change-Id: I4f57c9202ed4d055066286f808369ec0ddddb7ea