Update to build using charmcraft
Due to a build problem with the reactive plugin, this change falls back
on overriding the steps and doing a manual build, but it also ensures
the CI system builds the charm using charmcraft.  Changes:

- add a build-requirements.txt
- modify charmcraft.yaml
- modify osci.yaml
    -> indicate build with charmcraft
- modify tox.ini
    -> tox -e build does charmcraft build/rename
    -> tox -e build-reactive does the reactive build
- modify bundles to use the <charm>.charm artifact in tests.
  and fix deprecation warning re: prefix
- tox inception to enable tox -e func-test in the CI

This change also switches away from directory backed OSD devices
in the test bundles, as they are not supported anymore.

Change-Id: I57d1b47afbbeef211bb777fdbd0b4a091a021c19
Co-authored-by: Aurelien Lourot <aurelien.lourot@canonical.com>